refactor: extract utils functions to own files
