+++ src/CuteChat.tsx
@@ -13,6 +13,7 @@ import { Alert, NativeScrollEvent } from 'react-native';
 import type { IMessage } from 'react-native-gifted-chat';
 import { GiftedChat, GiftedChatProps } from 'react-native-gifted-chat';
 import { appendSnapshot } from './utils/appendSnapshot';
+import { prepareSnapshot } from './utils/prepareSnapshot';
 
 interface CustomCuteChatProps {
   chatId: string;
@@ -311,84 +312,3 @@ function isCloseToBottom({
     contentSize.height - paddingToBottom
   );
 }
-
-// Utility function to convert a Firestore document to a Gifted Chat message
-export const docToMessage = async (
-  doc: FirebaseFirestore.QueryDocumentSnapshot,
-  chatId: string
-): Promise<IMessage> => {
-  const data = doc.data();
-
-  if (!data) {
-    throw new Error('Document data is undefined');
-  }
-
-  const [files, sender] = await Promise.all([
-    new Promise<
-      | FirebaseFirestore.QueryDocumentSnapshot<FirebaseFirestore.DocumentData>[]
-      | undefined
-    >((resolve, reject) => {
-      firestore()
-        .collection(`chats/${chatId}/messages/${doc.id}/files`)
-        .onSnapshot((snapshot) => {
-          if (snapshot.empty) {
-            resolve(undefined);
-          } else {
-            resolve(snapshot.docs);
-          }
-        }, reject);
-    }),
-    new Promise<FirebaseFirestore.DocumentData | undefined>(
-      (resolve, reject) => {
-        firestore()
-          .doc(data.senderRef._documentPath._parts.join('/'))
-          .onSnapshot((snapshot) => {
-            if (!snapshot.exists) {
-              resolve(undefined);
-            } else {
-              resolve(snapshot.data());
-            }
-          }, reject);
-      }
-    ),
-  ]);
-
-  const image = files?.[0]?.data().url;
-
-  // Fetch user data from reference
-  if (sender) {
-    return {
-      _id: doc.id,
-      createdAt: new Date(data.createdAt),
-      text: data.content,
-      user: { _id: data.senderId, ...sender },
-      image: image,
-      readByIds: data.readByIds,
-      metadata: data.metadata,
-    };
-  } else {
-    return {
-      _id: doc.id,
-      createdAt: new Date(data.createdAt),
-      text: data.content,
-      image: image,
-      system: true,
-      readByIds: data.readByIds,
-      metadata: data.metadata,
-    };
-  }
-};
-
-export const prepareSnapshot = async (
-  snapshot: FirebaseFirestore.QuerySnapshot,
-  chatId: string
-): Promise<SnapshotChange[]> => {
-  console.log('Preparing snapshot');
-
-  return Promise.all(
-    snapshot.docChanges().map(async (change) => ({
-      type: change.type,
-      message: await docToMessage(change.doc, chatId),
-    }))
-  );
-};
